Trait data_model::volatile_memory::VolatileMemory

pub trait VolatileMemory {
    fn get_slice(
        &self, 
        offset: usize, 
        count: usize
    ) -> VolatileMemoryResult<VolatileSlice<'_>>;

    fn get_ref<T: DataInit>(
        &self, 
        offset: usize
    ) -> VolatileMemoryResult<VolatileRef<'_, T>> { ... }
}

Trait for types that support raw volatile access to their data.

Required methods

fn get_slice(
&self,
offset: usize,
count: usize
) -> VolatileMemoryResult<VolatileSlice<'_>>

Gets a slice of memory at offset that is count bytes in length and supports volatile access.

Provided methods

fn get_ref<T: DataInit>(
&self,
offset: usize
) -> VolatileMemoryResult<VolatileRef<'_, T>>

Gets a VolatileRef at offset.